Split config module's init into separate files
There's loggging and directory setup code in config's __init__.py that
doesn't belong there and leads to circular deps when trying to convert
it into the new setting access. Factor it out into new files and perform
setup in __main__.py.

This finishes preparations for creation of the new constant system.
